renamed build.ts to platform.ts
