+++ packages/cli/src/tokens/process/configs.ts
@@ -5,9 +5,9 @@ import type { Config as StyleDictionaryConfig, TransformedToken } from 'style-di
 import { outputReferencesFilter } from 'style-dictionary/utils';
 
 import { isColorCategoryToken, isDigit, isSemanticColorToken, pathStartsWithOneOf, typeEquals } from '../utils.js';
-import { buildOptions } from './build.js';
 import { formats } from './formats/css.js';
 import { jsTokens } from './formats/js-tokens.js';
+import { buildOptions } from './platform.js';
 import { resolveMath, sizeRem, typographyName, unitless } from './transformers.js';
 import type {
   BuiltInColors,
