+++ packages/cli/src/tokens/process/utils/getMultidimensionalThemes.ts
@@ -3,7 +3,7 @@ import { TokenSetStatus } from '@tokens-studio/types';
 import chalk from 'chalk';
 import { kebabCase } from 'change-case';
 import * as R from 'ramda';
-import { buildOptions } from '../build.js';
+import { buildOptions } from '../platform.js';
 import type { ThemeDimension, ThemePermutation } from '../types.js';
 
 /**
